chore: api connector domain re-organized
