+++ app/ui/src/app/customizations/api-connector/api-connector-detail/api-connector-detail.component.scss
