+++ app/ui/src/app/customizations/api-connector/api-connector-detail/api-connector-detail.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { ApiConnectorData } from './api-connector.models';
-import { ApiConnectorService } from './api-connector.service';
+import { ApiConnectorValidation } from '@syndesis/ui/customizations/api-connector';
+import { ApiConnectorService } from './../api-connector.service';
 
 @Component({
   selector: 'syndesis-api-connector-detail',
@@ -10,11 +10,13 @@ import { ApiConnectorService } from './api-connector.service';
   styleUrls: ['api-connector-detail.component.scss']
 })
 export class ApiConnectorDetailComponent implements OnInit {
-  apiConnectorData: ApiConnectorData;
+  apiConnectorData: ApiConnectorValidation;
   loading: boolean;
 
-  constructor(private apiConnectorService: ApiConnectorService,
-              private route: ActivatedRoute) {
+  constructor(
+    private apiConnectorService: ApiConnectorService,
+    private route: ActivatedRoute,
+  ) {
     this.loading = true;
   }
