+++ app/ui/src/app/customizations/api-connector/api-connector-detail/index.ts
@@ -1,0 +1,1 @@
+export * from './api-connector-detail.component';
