+++ app/ui/src/app/customizations/api-connector/api-connector-list/index.ts
@@ -1,0 +1,1 @@
+export * from './api-connector-list.component';
