+++ app/ui/src/app/customizations/api-connector/index.ts
@@ -1,6 +1,5 @@
+export * from './api-connector-list';
 export * from './api-connector.actions';
 export * from './api-connector.reducer';
 export * from './api-connector.models';
 export * from './api-connector.module';
-export * from './api-connector-list.component';
-export * from './api-connector-detail.component';
